// schemaregistry/confluent/type/variant_pb.ts
// @generated by protoc-gen-es v2.12.0 with parameter "target=ts"
// @generated from file confluent/type/variant.proto (package confluent.type, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file confluent/type/variant.proto.
 */
export const file_confluent_type_variant: GenFile = /*@__PURE__*/
  fileDesc("Chxjb25mbHVlbnQvdHlwZS92YXJpYW50LnByb3RvEg5jb25mbHVlbnQudHlwZSIqCgdWYXJpYW50EhAKCG1ldGFkYXRhGAEgASgMEg0KBXZhbHVlGAIgASgMYgZwcm90bzM");

/**
 * @generated from message confluent.type.Variant
 */
export type Variant = Message<"confluent.type.Variant"> & {
  /**
   * A dictionary of field names used by all objects in the variant value.
   * Encoded as: version header, dictionary size, offset list, and UTF-8 string data.
   *
   * @generated from field: bytes metadata = 1;
   */
  metadata: Uint8Array;

  /**
   * The variant value, which can be a primitive, object, or array.
   * Encoded as: a one-byte header (basic type + type info) followed by the content bytes.
   *
   * @generated from field: bytes value = 2;
   */
  value: Uint8Array;
};

/**
 * Describes the message confluent.type.Variant.
 * Use `create(VariantSchema)` to create a new message.
 */
export const VariantSchema: GenMessage<Variant> = /*@__PURE__*/
  messageDesc(file_confluent_type_variant, 0);